refactor the test and game stats to use the new game objects
